# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express MySQL Docker API",
    "version": "1.0.0",
    "description": "ノート・クイズ管理アプリケーションの REST API"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API サーバー"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT トークンを Authorization ヘッダーに設定"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "ログイン",
        "description": "メールアドレスとパスワードでログインし、JWTトークンを取得します。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "メールアドレス"
                  },
                  "password": {
                    "type": "string",
                    "description": "パスワード"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ログイン成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "JWTトークン"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "リクエストパラメータ不正",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "メールアドレスとパスワードは必須です"
                }
              }
            }
          },
          "401": {
            "description": "認証失敗",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "メールアドレスまたはパスワードが正しくありません"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "ログインユーザー情報取得",
        "description": "認証済みユーザーの情報を取得します。Bearerトークンが必要です。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "取得成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    },
                    "role": {
                      "type": "string",
                      "description": "ユーザーロール"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "ユーザーが見つからない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "ユーザーが見つかりません"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Auth"
        ],
        "summary": "ログインユーザー情報更新",
        "description": "認証済みユーザーの情報を更新します。Bearerトークンが必要です。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "名前"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "メールアドレス"
                  },
                  "currentPassword": {
                    "type": "string",
                    "description": "現在のパスワード（パスワード変更時のみ必須）"
                  },
                  "newPassword": {
                    "type": "string",
                    "description": "新しいパスワード（6文字以上）"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "現在のパスワードが正しくない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "メールアドレス重複",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "このメールアドレスはすでに使用されています"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}